Docs: Improve the phrasing in the Game Mechanics intro
This one's weird: I think the original phrasing is redundant since the same information is already present on the About page (currently a draft, but still). Also, it's not really the goal of the Game Mechanics category, but rather part of the goals for the entire project, and as such it doesn't belong on this page.