Docs: Add placeholder categories for otherwise dead links
The about page references the two major categories that I've yet to create, so the CI will fail until something is put in their place. While this isn't ideal, it won't take long until the Rendering parts are there (since they're part of the file formats repository currently in the process of being migrated), and Arcturus... well, it's not exactly important so I doubt anyone cares.