Docs: Rework the SPR format specification with diff highlights
This cuts down a lot on redundancy and verbosity, while hopefully still providing all of the same information. It's mostly a test for the diff views; if this works out then the other format specifications can be reworked as well. Also added some reference implementations that are reasonably easy to understand (omitting the Unity ones because they aren't suitable for this purpose).